services/scanner/requirements-scanner.txt
websockets==12.0
pandas==2.2.0
numpy==1.26.0
redis[hiredis]==5.0.1
python-dotenv==1.0.0
requests==2.31.0
pika==1.3.2
boto3==1.34.0
aiohttp==3.9.1
pybit==5.7.0
